Fixed an overseen problem
Fixed a problem where text from a file was being indexed and put into not just its own dictionary but the dictionary of the files that get indexed after it.
dictionaries in diction_list in Indexing.py held tokens of other files (specifically previously indexed files) in them instead of just their own.